+++ src/lib/validations/auth.ts
@@ -20,9 +20,6 @@ export const LoginSchema = z.object({
 
 export const SignupSchema = z
   .object({
-    name: z.string().min(3, {
-      message: "Name must be at least 3 characters long.",
-    }),
     username: z
       .string()
       .min(3, { message: "Username must be at least 3 characters long." })